(engine) fix THREAD_T and THREAD_ID definition when threads are disabled
Src-commit: ed5db8f96f2006fff645d43874ee716b446f32ac

+++ core/engine/os_threads.h
@@ -575,8 +575,8 @@ typedef LPTHREAD_START_ROUTINE  THREAD_START; /* The type of the routine */
 #include <sys/wait.h>
 #endif
 
-typedef int THREAD_T;
-typedef int THREAD_ID;  /* The unique identifier of a thread */
+typedef intptr_t THREAD_T;
+typedef intptr_t THREAD_ID;  /* The unique identifier of a thread */
 typedef void *THREAD_ARG;
 typedef void *THREAD_RES_T;
 typedef void *(*THREAD_START)(void *);
